BIGINT -> INT

+++ src/main/resources/db/migration/V1__create_table_soknad.sql
@@ -1,6 +1,6 @@
 CREATE TABLE SOKNAD
 (
-    id                 BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
+    id                 INT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
     uuid               UUID        NOT NULL UNIQUE,
     ekstern_id         UUID        NOT NULL UNIQUE,
     personident        VARCHAR(11) NOT NULL,
@@ -11,8 +11,8 @@ CREATE INDEX idx_soknad_personident ON SOKNAD (personident);
 
 CREATE TABLE SOKNAD_PERIODE
 (
-    id        BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
-    soknad_id BIGINT NOT NULL REFERENCES SOKNAD (id) ON DELETE CASCADE,
+    id        INT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
+    soknad_id INT NOT NULL REFERENCES SOKNAD (id) ON DELETE CASCADE,
     fom       DATE   NOT NULL,
     tom       DATE   NOT NULL,
     CONSTRAINT check_soknad_periode_fom_tom CHECK (fom <= tom)
@@ -23,9 +23,9 @@ CREATE TYPE VEDTAK_UTFALL AS ENUM ('INNVILGET');
 
 CREATE TABLE VEDTAK
 (
-    id         BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
+    id         INT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
     uuid       UUID          NOT NULL UNIQUE,
-    soknad_id  BIGINT        NOT NULL REFERENCES SOKNAD (id) ON DELETE CASCADE,
+    soknad_id  INT        NOT NULL REFERENCES SOKNAD (id) ON DELETE CASCADE,
     utfall     VEDTAK_UTFALL NOT NULL,
     fattet_av  VARCHAR(30)   NOT NULL,
     created_at TIMESTAMPTZ   NOT NULL DEFAULT NOW()
@@ -34,8 +34,8 @@ CREATE INDEX idx_vedtak_soknad_id ON VEDTAK (soknad_id);
 
 CREATE TABLE VEDTAK_PERIODE
 (
-    id        BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
-    vedtak_id BIGINT NOT NULL REFERENCES VEDTAK (id) ON DELETE CASCADE,
+    id        INT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
+    vedtak_id INT NOT NULL REFERENCES VEDTAK (id) ON DELETE CASCADE,
     fom       DATE   NOT NULL,
     tom       DATE   NOT NULL,
     CONSTRAINT check_vedtak_periode_fom_tom CHECK (fom <= tom)
